Refactor Garmin Connect API calls into one class
