+++ garmin_workouts/import_workouts.py
@@ -2,19 +2,17 @@ import logging
 import json
 import yaml
 
-from login import Login
 from libs.parser import WorkoutParser
+from libs.garmin_api_client import GarminApiClient
 
 log = logging.getLogger(__name__)
 
 
 class Import():
-    def __init__(self, args):
-        login = Login()
-        login.login()
-        self.session = login.get_session()
-        self.filename = args.import_file
+    def __init__(self, args, session):
 
+        self.filename = args.import_file
+        self.api_client = GarminApiClient(session=session)
         self.import_workouts()
 
     def import_workouts(self):
@@ -22,33 +20,11 @@ class Import():
 
         workout_parser = WorkoutParser(own_format=workout_obj)
         garmin_workout = workout_parser.get_garmin_format()
+        workout_name = workout_parser.get_workout_name()
 
         workout_json = json.dumps(garmin_workout, sort_keys=True, indent=2)
 
-        assert self.session
-
-        log.info(f"Importing a workout from file {self.filename} to the Garmin Connect")
-        headers = {
-            "Accept": "application/json, text/javascript, */*; q=0.01",
-            "Accept-Language": "en-US,en;q=0.5",
-            "NK": "NT",
-            "Referer": "https://connect.garmin.com/modern/workouts",
-            "Content-Type": "application/json",
-        }
-        response = self.session.post(
-            "https://connect.garmin.com/modern/proxy/workout-service/workout",
-            headers=headers, data=workout_json)
-        response.raise_for_status()
-
-        response_json = response.json()
-        new_workout_url = f"https://connect.garmin.com/modern/workout/{response_json['workoutId']}"
-        log.info(f'New workout created: {new_workout_url}')
-
-        # if self.keep:
-        #     with open(self.response_filename, 'w') as outfile:
-        #         log.debug('Storing response to the %s'
-        #                     % self.response_filename)
-        #         json.dump(response_json, outfile, indent=4)
+        self.api_client.upload_new_workout(workout_json, workout_name)
 
     def parse_yaml_file(self, filename):
         with open(filename, 'r') as infile:
